Switch to non-perfect cubes as cells in LinkedCellParticleContainer (needed to handle flexible domain sizes) and add documentation to all new funcions

+++ src/utils/Generator.h
@@ -8,13 +8,44 @@
 #include "models/ParticleContainer.h"
 
 namespace Generator {
-    void
-    cuboid(ParticleContainer &container, std::array<double, 3> position, std::array<int, 3> size, double meshWidth,
+    /**
+     * @brief Generates a cuboid of particles
+     *
+     * @param container
+     * @param position
+     * @param size
+     * @param meshWidth
+     * @param velocity
+     * @param mass
+     * @param typeId
+     */
+    void cuboid(ParticleContainer &container, std::array<double, 3> position, std::array<double, 3> size, double meshWidth,
            std::array<double, 3> velocity, double mass, int typeId);
 
+    /**
+     * @brief Generates a sphere of particles
+     * @param container
+     * @param center
+     * @param radius
+     * @param meshWidth
+     * @param velocity
+     * @param mass
+     * @param typeId
+     */
     void sphere(ParticleContainer &container, std::array<double, 3> center, int radius, double meshWidth,
                 std::array<double, 3> velocity, double mass, int typeId);
 
+    /**
+     * @brief Generates a disk of particles
+     * @note Equivalent to a '2d sphere'
+     * @param container
+     * @param center
+     * @param radius
+     * @param meshWidth
+     * @param velocity
+     * @param mass
+     * @param typeId
+     */
     void disk(ParticleContainer &container, std::array<double, 3> center, int radius, double meshWidth,
                 std::array<double, 3> velocity, double mass, int typeId);
